+++ src/utils/sort.ts
@@ -54,21 +54,32 @@ export function orderByComparator(a: any, b: any): number {
 
 /**
  * Sorts the rows
- * @param  {Array<any>}  rows
- * @param  {Array<Sort>} dirs
- * @return {Array<any>} results
+ * 
+ * @export
+ * @param {any[]} rows
+ * @param {any[]} columns
+ * @param {any[]} dirs
+ * @returns
  */
-export function sortRows(rows: Array<any>, dirs: any[]) {
-  let temp = [...rows];
+export function sortRows(rows: any[], columns: any[], dirs: any[]) {
+  if(!rows || !dirs || !columns) return rows;
 
-  return temp.sort(function(a: any, b: any) {
+  let cols = columns.reduce((obj, col) => {
+    if(col.comparator && typeof col.comparator === 'function') {
+      obj[col.prop] = col.comparator;
+    }
+    return obj;
+  }, {});
+
+  return [...rows].sort(function(a: any, b: any) {
     for(const { prop, dir } of dirs) {
-      let propA = deepValueGetter(a, prop);
-      let propB = deepValueGetter(b, prop);
+      const propA = deepValueGetter(a, prop);
+      const propB = deepValueGetter(b, prop);
 
+      const compareFn = cols[prop] || orderByComparator;
       const comparison = dir !== SortDirection.desc ?
-        orderByComparator(propA, propB) :
-        -orderByComparator(propA, propB);
+        compareFn(propA, propB) :
+        -compareFn(propA, propB);
 
       // Don't return 0 yet in case of needing to sort by next property
       if (comparison !== 0) return comparison;